requirements.txt
flask==2.0.3
flask-sqlalchemy==3.0.3
flask-login==0.6.2
flask-wtf==1.0.0
flask-migrate==4.0.4
flask-socketio==5.3.4
python-dotenv==1.0.0
python-engineio==4.3.4
python-socketio==5.7.2
werkzeug==2.0.3
wtforms==3.0.1